% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgat_function.r
\name{cgat}
\alias{cgat}
\title{Chaos Game Representation (CGR) Frequency Matrix}
\usage{
cgat(k_mer, seq_number, len_trim, fasta_filtered)
}
\arguments{
\item{k_mer}{Integer. The k-mer size used to define resolution (e.g., 6 for 4^6 bins).}

\item{seq_number}{Integer. The index of the sequence in \code{fasta_filtered} to process.}

\item{len_trim}{Integer. The number of bases to use from the start of the sequence.}

\item{fasta_filtered}{A list of filtered DNA sequences.}
}
\value{
A square numeric matrix of size sqrt(4^k_mer) x sqrt(4^k_mer) representing the frequency of k-mers at each CGR coordinate.
}
\description{
Generates a CGR frequency matrix from a given DNA sequence using a specified k-mer size.
}
\details{
This function uses a recursive approach to generate a 2D CGR matrix from DNA. It assumes that a global object named \code{fasta_filtered} is present and contains DNA sequences accessible by index.
}
\examples{
# Example assuming fasta_filtered is loaded:
# fasta_filtered <- list("ATCGTAGCTAAGCTAGCT")
# cgr_mat <- cgat(k_mer = 6, seq_number = 1, len_trim = 100)

}
