% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgrplot.r
\name{cgrplot}
\alias{cgrplot}
\title{Chaos Game Representation (CGR) Coordinate Matrix}
\usage{
cgrplot(seq_number, fasta_filtered)
}
\arguments{
\item{seq_number}{Integer. The index of the sequence in \code{fasta_filtered} to process.}

\item{fasta_filtered}{A list of filtered DNA sequences.}
}
\value{
A numeric matrix with two columns and one row per base. The columns represent the X and Y coordinates of each base in the CGR.
}
\description{
Computes the CGR 2D coordinates for a DNA sequence without binning, for plotting purposes.
}
\details{
This function generates the sequence of (x, y) coordinates used to construct a Chaos Game Representation (CGR) plot of a DNA sequence. It assumes a global object named \code{fasta_filtered} contains the DNA sequences.
}
\examples{
fasta_filtered <- fastafile_new(read.fasta("example.fasta"), 50)
cgrplot(1, fasta_filtered)

}
